+++ Vector3.h
@@ -11,21 +11,21 @@ class Vector3 {
 public:
     Vector3();
 
-    Vector3(float x, float y, float z);
+    Vector3(double x, double y, double z);
 
-    float x, y, z;
+    double x, y, z;
 
     static Vector3 vectorFromTwoPoints(Vector3 point1, Vector3 point2);
 
     Vector3 add(Vector3 vector2);
 
-    Vector3 divideByNumber(float number);
+    Vector3 divideByNumber(double number);
 
     Vector3 multiply(Vector3 vector2);
 
     Vector3 normalize();
 
-    static float dot(float x1, float y1, float z1, Vector3 vector);
+    static double dot(double x1, double y1, double z1, Vector3 vector);
 
     static Vector3 cross(Vector3 vector1, Vector3 vector2);
